Rename async to async_helpers because `async` is a keyword on Python 3.7 (#3678)
